Reorganize core handler structure
Consolidate handler files and move handler families into dedicated namespaces/folders.


Copilot-Session: 1a86824c-be19-4cde-a42b-dff6f40b8189
