+++ core/samples/MessageExtensionBot/Program.cs
@@ -4,7 +4,7 @@
 using System.Text.Json;
 using MessageExtensionBot;
 using Microsoft.Teams.Apps;
-using Microsoft.Teams.Apps.MessageExtension;
+using Microsoft.Teams.Apps.Handlers.MessagingExtensions;
 using Microsoft.Teams.Apps.Schema;
 using Microsoft.Teams.Apps.TaskModules;
