+++ core/src/Microsoft.Teams.Apps/Handlers/AgentLifecycleHandlers.Activity.cs
