+++ core/src/Microsoft.Teams.Apps/Handlers/AgentLifecycleHandlers.Values.cs
