+++ core/src/Microsoft.Teams.Apps/Handlers/AgentLifecycleHandlers.cs
