+++ core/src/Microsoft.Teams.Apps/Handlers/ConversationUpdateHandlers.Activity.cs
