+++ core/src/Microsoft.Teams.Apps/Handlers/ConversationUpdateHandlers.cs
