+++ core/src/Microsoft.Teams.Apps/Handlers/Files/FileConsentHandlers.Values.cs
@@ -3,7 +3,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace Microsoft.Teams.Apps;
+namespace Microsoft.Teams.Apps.Files;
 
 /// <summary>
 /// Represents the value of the invoke activity sent when the user acts on a
