+++ core/src/Microsoft.Teams.Apps/Handlers/Files/FileConsentHandlers.cs
@@ -3,24 +3,26 @@
 
 using Microsoft.Teams.Apps.Routing;
 using Microsoft.Teams.Apps.Schema;
+using Microsoft.Teams.Apps;
 
-namespace Microsoft.Teams.Apps;
+namespace Microsoft.Teams.Apps.Files;
 
 /// <summary>
-/// Delegate for handling adaptive card action invoke activities.
+/// Delegate for handling file consent invoke activities.
 /// </summary>
 /// <param name="context">The context for the invoke activity, providing access to the activity data and bot application.</param>
 /// <param name="cancellationToken">A cancellation token that can be used to cancel the operation.</param>
 /// <returns>A task that represents the asynchronous operation. The task result contains the invoke response.</returns>
-public delegate Task<InvokeResponse> AdaptiveCardActionHandler(Context<InvokeActivity<AdaptiveCardActionValue>> context, CancellationToken cancellationToken = default);
+public delegate Task<InvokeResponse<AdaptiveCardResponse>> FileConsentValueHandler(Context<InvokeActivity<FileConsentValue>> context, CancellationToken cancellationToken = default);
 
 /// <summary>
-/// Extension methods for registering adaptive card action invoke handlers.
+/// Extension methods for registering file consent invoke handlers.
 /// </summary>
-public static class AdaptiveCardExtensions
+public static class FileConsentExtensions
 {
+
     /// <summary>
-    /// Registers a handler for adaptive card action invoke activities.
+    /// Registers a handler for file consent invoke activities.
     /// Cannot be combined with <see cref="InvokeExtensions.OnInvoke"/>.
     /// </summary>
     /// <remarks>
@@ -29,16 +31,16 @@ public static class AdaptiveCardExtensions
     /// <param name="app">The Teams bot application.</param>
     /// <param name="handler">The handler to register.</param>
     /// <returns>The updated Teams bot application.</returns>
-    public static TeamsBotApplication OnAdaptiveCardAction(this TeamsBotApplication app, AdaptiveCardActionHandler handler)
+    public static TeamsBotApplication OnFileConsent(this TeamsBotApplication app, FileConsentValueHandler handler)
     {
         ArgumentNullException.ThrowIfNull(app, nameof(app));
         app.Router.Register(new Route<InvokeActivity>
         {
-            Name = string.Join("/", TeamsActivityTypes.Invoke, InvokeNames.AdaptiveCardAction),
-            Selector = activity => activity.Name == InvokeNames.AdaptiveCardAction,
+            Name = string.Join("/", TeamsActivityTypes.Invoke, InvokeNames.FileConsent),
+            Selector = activity => activity.Name == InvokeNames.FileConsent,
             HandlerWithReturn = async (ctx, cancellationToken) =>
             {
-                InvokeActivity<AdaptiveCardActionValue> typedActivity = new(ctx.Activity);
+                InvokeActivity<FileConsentValue> typedActivity = new(ctx.Activity);
                 var typedContext = ctx.CreateDerivedContext(typedActivity);
                 return await handler(typedContext, cancellationToken).ConfigureAwait(false);
             }
