+++ core/src/Microsoft.Teams.Apps/Handlers/InstallUpdateHandlers.Activity.cs
