+++ core/src/Microsoft.Teams.Apps/Handlers/InstallUpdateHandlers.cs
