+++ core/src/Microsoft.Teams.Apps/Handlers/Meetings/MeetingHandlers.Values.cs
@@ -4,7 +4,7 @@
 using System.Text.Json.Serialization;
 using Microsoft.Teams.Apps.Schema;
 
-namespace Microsoft.Teams.Apps;
+namespace Microsoft.Teams.Apps.Meetings;
 
 /// <summary>
 /// Value payload for a meeting start event.
