+++ core/src/Microsoft.Teams.Apps/Handlers/MessageExtensions/MessageExtensionActionResponse.cs
@@ -4,7 +4,7 @@
 using System.Text.Json.Serialization;
 using Microsoft.Teams.Apps.TaskModules;
 
-namespace Microsoft.Teams.Apps.MessageExtension;
+namespace Microsoft.Teams.Apps.Handlers.MessagingExtensions;
 
 /// <summary>
 /// Represents a response from a message extension action that can contain either a task module or compose extension response.
