+++ core/src/Microsoft.Teams.Apps/Handlers/MessageExtensions/MessageExtensionHandlers.cs
@@ -5,7 +5,7 @@ using System.Text.Json;
 using Microsoft.Teams.Apps.Routing;
 using Microsoft.Teams.Apps.Schema;
 
-namespace Microsoft.Teams.Apps.MessageExtension;
+namespace Microsoft.Teams.Apps.Handlers.MessagingExtensions;
 
 /// <summary>
 /// Delegate for handling message extension query invoke activities.
