+++ core/src/Microsoft.Teams.Apps/Handlers/MessageExtensions/MessageExtensionQueryLink.cs
@@ -3,7 +3,7 @@
 
 using System.Text.Json.Serialization;
 
-namespace Microsoft.Teams.Apps.MessageExtension;
+namespace Microsoft.Teams.Apps.Handlers.MessagingExtensions;
 
 /// <summary>
 /// App-based query link payload for link unfurling.
