+++ core/src/Microsoft.Teams.Apps/Handlers/MessageExtensions/MessageExtensionResponse.cs
@@ -5,7 +5,7 @@ using System.Text.Json.Serialization;
 using Microsoft.Teams.Apps.Schema;
 using Microsoft.Teams.Apps.Utils;
 
-namespace Microsoft.Teams.Apps.MessageExtension;
+namespace Microsoft.Teams.Apps.Handlers.MessagingExtensions;
 
 /// <summary>
 /// Messaging extension response types.
