+++ core/src/Microsoft.Teams.Apps/Handlers/MessageHandlers.cs
@@ -16,6 +16,22 @@ namespace Microsoft.Teams.Apps;
 public delegate Task MessageHandler(Context<MessageActivity> context, CancellationToken cancellationToken = default);
 
 /// <summary>
+/// Delegate for handling message update activities.
+/// </summary>
+/// <param name="context">The context for the message update activity.</param>
+/// <param name="cancellationToken">A cancellation token that can be used to cancel the operation.</param>
+/// <returns>A task representing the asynchronous operation.</returns>
+public delegate Task MessageUpdateHandler(Context<MessageUpdateActivity> context, CancellationToken cancellationToken = default);
+
+/// <summary>
+/// Delegate for handling message delete activities.
+/// </summary>
+/// <param name="context">The context for the message delete activity.</param>
+/// <param name="cancellationToken">A cancellation token that can be used to cancel the operation.</param>
+/// <returns>A task representing the asynchronous operation.</returns>
+public delegate Task MessageDeleteHandler(Context<MessageDeleteActivity> context, CancellationToken cancellationToken = default);
+
+/// <summary>
 /// Extension methods for registering message activity handlers.
 /// </summary>
 public static class MessageExtensions
@@ -100,5 +116,54 @@ public static class MessageExtensions
 
         return app;
     }
-}
 
+    /// <summary>
+    /// Registers a handler for message update activities.
+    /// </summary>
+    /// <remarks>
+    /// Breaking change: previously only the first matching handler was invoked. All matching handlers are now invoked sequentially.
+    /// </remarks>
+    /// <param name="app">The Teams bot application.</param>
+    /// <param name="handler">The handler to register.</param>
+    /// <returns>The updated Teams bot application.</returns>
+    public static TeamsBotApplication OnMessageUpdate(this TeamsBotApplication app, MessageUpdateHandler handler)
+    {
+        ArgumentNullException.ThrowIfNull(app, nameof(app));
+        app.Router.Register(new Route<MessageUpdateActivity>
+        {
+            Name = TeamsActivityTypes.MessageUpdate,
+            Selector = _ => true,
+            Handler = async (ctx, cancellationToken) =>
+            {
+                await handler(ctx, cancellationToken).ConfigureAwait(false);
+            }
+        });
+
+        return app;
+    }
+
+    /// <summary>
+    /// Registers a handler for message delete activities.
+    /// </summary>
+    /// <remarks>
+    /// Breaking change: previously only the first matching handler was invoked. All matching handlers are now invoked sequentially.
+    /// </remarks>
+    /// <param name="app">The Teams bot application.</param>
+    /// <param name="handler">The handler to register.</param>
+    /// <returns>The updated Teams bot application.</returns>
+    public static TeamsBotApplication OnMessageDelete(this TeamsBotApplication app, MessageDeleteHandler handler)
+    {
+        ArgumentNullException.ThrowIfNull(app, nameof(app));
+        app.Router.Register(new Route<MessageDeleteActivity>
+        {
+            Name = TeamsActivityTypes.MessageDelete,
+            Selector = _ => true,
+            Handler = async (ctx, cancellationToken) =>
+            {
+                await handler(ctx, cancellationToken).ConfigureAwait(false);
+            }
+        });
+
+        return app;
+    }
+}
