+++ core/test/Microsoft.Teams.Apps.UnitTests/HandlerStatePropagationTests.cs
@@ -6,8 +6,10 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Teams.Apps;
+using Microsoft.Teams.Apps.Files;
+using Microsoft.Teams.Apps.Meetings;
 using Microsoft.Teams.Apps.Clients;
-using Microsoft.Teams.Apps.MessageExtension;
+using Microsoft.Teams.Apps.Handlers.MessagingExtensions;
 using Microsoft.Teams.Apps.Schema;
 using Microsoft.Teams.Apps.State;
 using Microsoft.Teams.Apps.TaskModules;
